requirements.txt
openai
python-dotenv
click
rich
tiktoken
pydantic